requirements.txt
streamlit
pyaudio
numpy
scipy
selenium
openai-whisper
